requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chatterbox-tts==0.1.3
click==8.1.8
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
groq==0.23.1
gTTS==2.5.4
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
reportlab
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.60
langchain-groq==0.3.2
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
Markdown==3.8
MarkupSafe==3.0.2
motor==3.3.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
narwhals==1.39.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==1.76.2
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
propcache==0.3.1
protobuf==6.30.2
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
pymongo==4.5.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
Resemblyzer==0.1.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.41
srt==3.5.3
starlette==0.46.2
streamlit==1.45.1
structlog==25.3.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
torch==2.6.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
watchdog==6.0.0
webrtcvad==2.0.10
websockets==15.0.1
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
mysql-connector-python
